🚑️ Hotfix:  리다이렉션 로직 다른파일로 이동
